requirements.txt
cupy-cuda12x
numpy
matplotlib
scipy
Pillow
